+++ doc/CHANGES.rst
@@ -18,7 +18,7 @@ Bugs Fixed
 Features Added
 ++++++++++++++
 
-- Added notification of IDatabaseOpenedWithRoot.
+- Added notification of ``IDatabaseOpenedWithRoot``.
 
 - Added a new API's ``get_registered_packages, set_registered_packages`` to
   ``OFS.metaconfigure`` which replace any direct access to