+++ doc/CHANGES.rst
@@ -11,6 +11,8 @@ http://docs.zope.org/zope2/releases/.
 Bugs Fixed
 ++++++++++
 
+- LP #143531: Fix broken object so they give access to their state.
+
 - LP #578326: Issue a warning if someone specifies a non-public permission
   attribute in the browser:view directive. This attribute has never been
   supported in Zope 2.